fix: Squash migrations and fix entity mappings

+++ backend/mercatus/src/main/java/com/scientia/mercatus/dto/CartItemDto.java
@@ -4,5 +4,5 @@ package com.scientia.mercatus.dto;
 
 import java.math.BigDecimal;
 
-public record CartItemDto(Long id, String name, BigDecimal quantity, BigDecimal unitPrice, BigDecimal totalItemsPrice) {
+public record CartItemDto(Long id, String name, Integer quantity, BigDecimal unitPrice, BigDecimal totalItemsPrice) {
 }
