+++ backend/mercatus/src/main/java/com/scientia/mercatus/dto/UpdateQuantityRequestDto.java
@@ -2,5 +2,5 @@ package com.scientia.mercatus.dto;
 
 import java.math.BigDecimal;
 
-public record UpdateQuantityRequestDto(Long productId, BigDecimal quantity) {
+public record UpdateQuantityRequestDto(Long productId, Integer quantity) {
 }
