+++ backend/mercatus/src/main/java/com/scientia/mercatus/service/ICartService.java
@@ -8,10 +8,10 @@ import java.math.BigDecimal;
 
 public interface ICartService {
     Cart resolveCart(CartContextDto cartContext);
-    void addToCart(Cart currentCart, Long productId, BigDecimal quantity);
+    void addToCart(Cart currentCart, Long productId, Integer quantity);
     void removeFromCart(Cart currentCart, Long productId);
     void clearCart(Cart currentCart);
-    void updateQuantity(Cart currentCart, Long productId, BigDecimal quantity);
+    void updateQuantity(Cart currentCart, Long productId, Integer quantity);
     CartResponseDto getCartDetails(Cart Cart);
 
 }
